refactor: move NumberRange to shared observability model package
NumberRange was in analytics_engine but also used by logs_engine.
Move it to io.gravitee.apim.core.observability.model alongside
the other shared types (FilterType, FilterOperator).
