+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/analytics_engine/model/TimeSeriesRequest.java
@@ -15,6 +15,7 @@
  */
 package io.gravitee.apim.core.analytics_engine.model;
 
+import io.gravitee.apim.core.observability.model.NumberRange;
 import java.util.ArrayList;
 import java.util.List;
