+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/apim/core/logs_engine/model/LogsFilterSpec.java
@@ -15,9 +15,9 @@
  */
 package io.gravitee.apim.core.logs_engine.model;
 
-import io.gravitee.apim.core.analytics_engine.model.NumberRange;
 import io.gravitee.apim.core.observability.model.FilterOperator;
 import io.gravitee.apim.core.observability.model.FilterType;
+import io.gravitee.apim.core.observability.model.NumberRange;
 import java.util.List;
 
 /**
